backend/pyproject.toml
[tool.poetry]
name = "fdyf-backend"
version = "0.1.0"
description = ""
authors = ["Jaeyeong Yang <jaeyeong.yang1125@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-filter = "^21.1"
jupyterlab = "^3.2.3"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pyflakes = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
